rename sparse_map3d
