+++ source/consumers/dataspaces/CMakeLists.txt
@@ -5,7 +5,7 @@ set(SOURCES
   ${dir}/dense_matrix2d.cpp
   ${dir}/scalar.cpp
   ${dir}/sparse_map2d.cpp
-  ${dir}/sparse_map3d.cpp
+  ${dir}/SparseMap3D.cpp
   ${dir}/SparseMatrix2D.cpp
   )
 
@@ -14,7 +14,7 @@ set(HEADERS
   ${dir}/dense_matrix2d.h
   ${dir}/scalar.h
   ${dir}/sparse_map2d.h
-  ${dir}/sparse_map3d.h
+  ${dir}/SparseMap3D.h
   ${dir}/SparseMatrix2D.h
   )
